refactor: move getLoginRedirect() and getLogoutRedirect() to Config\Auth
